[CHORE] 기본 투두 아이템 원복
Fixes: TICO-44
Related to: TICO-33
