+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/todo/view/TodoListScreen.kt
@@ -33,6 +33,8 @@ import com.tico.pomorodo.ui.common.view.SimpleText
 import com.tico.pomorodo.ui.iconpack.commonIconPack.IcFavoriteFilled
 import com.tico.pomorodo.ui.iconpack.commonIconPack.IcGroup
 import com.tico.pomorodo.ui.theme.IC_ADD_TODO
+import com.tico.pomorodo.ui.theme.IC_TODO_CHECKED
+import com.tico.pomorodo.ui.theme.IC_TODO_GOING
 import com.tico.pomorodo.ui.theme.IC_TODO_MORE_INFO
 import com.tico.pomorodo.ui.theme.IC_TODO_UNCHECKED
 import com.tico.pomorodo.ui.theme.IconPack
@@ -189,8 +191,13 @@ fun TodoListItem(
         horizontalArrangement = Arrangement.spacedBy(5.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
-        BasicTodoItem(todoData = todoData, onStateChanged = onStateChanged)
-
+        TodoCheckBox(state = todoData.state, onStateChanged = onStateChanged)
+        SimpleText(
+            modifier = Modifier.weight(1f),
+            text = todoData.name,
+            style = PomoroDoTheme.typography.laundryGothicRegular14,
+            color = PomoroDoTheme.colorScheme.onBackground
+        )
         if (isGroup) {
             Column(
                 verticalArrangement = Arrangement.spacedBy(5.dp),
@@ -283,4 +290,28 @@ fun TodoItem(
             color = PomoroDoTheme.colorScheme.onBackground
         )
     }
+}
+
+@Composable
+private fun TodoCheckBox(state: TodoState, onStateChanged: (TodoState) -> Unit) {
+    SimpleIcon(
+        modifier = Modifier.clickable {
+            onStateChanged(state)
+        },
+        size = 26,
+        imageVector = when (state) {
+            TodoState.UNCHECKED -> PomoroDoTheme.iconPack[IC_TODO_UNCHECKED]!!
+            TodoState.GOING -> PomoroDoTheme.iconPack[IC_TODO_GOING]!!
+            TodoState.CHECKED -> PomoroDoTheme.iconPack[IC_TODO_CHECKED]!!
+        },
+        contentDescriptionId = when (state) {
+            TodoState.UNCHECKED -> R.string.content_todo_unchecked
+            TodoState.GOING -> R.string.content_todo_going
+            TodoState.CHECKED -> R.string.content_todo_checked
+        }
+    )
+}
+
+enum class TodoState {
+    UNCHECKED, GOING, CHECKED
 }
